feat:用户对话，与自己对话 用户置顶优化 消息入库

+++ tudoIM_chat/chat_rpc/internal/logic/userchatlogic.go
@@ -44,7 +44,7 @@ func (l *UserChatLogic) UserChat(in *chat_rpc.UserChatRequest) (*chat_rpc.UserCh
 	chat := chat_models.ChatModel{
 		SendUserID: uint(in.SendUserId),
 		RevUserID:  uint(in.RevUserId),
-		MsgType:    int8(msg.Type),
+		MsgType:    msg.Type,
 		Msg:        msg,
 		SystemMsg:  systemMsg,
 	}
